Implement seperation of concerns for vue components and views
