+++ src/components/GlobalSelectionSwitch/globalSelectionSwitch.scss
@@ -1,0 +1,48 @@
+@import "../../design/main";
+
+.perk-switch-container {
+  position: relative;
+  margin: 0 auto;
+  width: var(--elementlength, 128px);
+  height: var(--elementlength, 128px);
+
+  .perk-switch {
+    position: absolute;
+    border: 1px solid rgba(255, 255, 255, 0.2);
+    background-color: rgba(0, 0, 0, 0.2);
+    background-image: url(/img/img_blood.png);
+    transition: all .15s ease-in-out;
+    transform: scale(1);
+    background-size: cover;
+    width: var(--elementlength, 128px);
+    height: var(--elementlength, 128px);
+
+    @media screen and (min-width: 992px) {
+      &:hover {
+        background-color: rgba(0, 0, 0, 0.4);
+      }
+    }
+
+    .perk-switch__name {
+      position: absolute;
+      top: var(--elementlength, 128px);
+      width: var(--elementlength, 128px);
+      height: 26px;
+    }
+
+    .perk-switch__image-container {
+      top: 0;
+      position: relative;
+      width: var(--elementlength, 128px);
+      height: var(--elementlength, 128px);
+    }
+
+    .perk-switch__image {
+      position: absolute;
+      top: 0;
+      width: var(--elementlength, 100px);
+      height: var(--elementlength, 100px);
+      background: var(--slotBg);
+    }
+  }
+}
