+++ src/components/InfoText/infoText.js
@@ -1,0 +1,10 @@
+export default {
+  name: 'InfoText',
+
+  props: {
+    value: {
+      required: false
+    }
+  }
+}
+
