+++ src/components/InfoText/infoText.scss
@@ -1,0 +1,45 @@
+@import "../../design/main";
+@import "../../design/typography";
+
+.paragraph {
+  padding: 10px;
+
+  a {
+    color: $color-primary;
+    text-decoration: none;
+  }
+
+  h3 {
+    margin-top: 10px
+  }
+
+  a:hover {
+    text-decoration: underline;
+  }
+
+  a:active {
+    color: black;
+  }
+
+  a:visited {
+    color: darkgrey;
+  }
+}
+
+li {
+  margin-left: 20px;
+}
+
+.video {
+  min-height: 315px;
+  width: 100%;
+  height: 100%;
+  max-width: 1024px;
+  margin: 0 auto;
+}
+
+.e-mail:before {
+  content: attr(data-website) "\0040" attr(data-user);
+  unicode-bidi: bidi-override;
+  direction: rtl;
+}
