+++ src/components/MenuItem/menuItem.scss
@@ -1,0 +1,63 @@
+@import "../../design/main";
+
+a {
+  color: white;
+}
+
+.perk-overview {
+  margin-bottom: 30px;
+
+  .perk-overview__toggle {
+    font-size: 25px;
+    font-weight: 500;
+    cursor: pointer;
+    text-transform: uppercase;
+
+    &:hover {
+      border: 1px solid rgba(255, 255, 255, 0.3);
+      background-color: rgba(0, 0, 0, 0.3);
+    }
+  }
+
+  .perk-overview__icon-holder {
+    display: block;
+    margin-right: 20px;
+    width: 120px;
+    flex-shrink: 0;
+
+    img {
+      display: block;
+      max-width: 100%;
+    }
+  }
+
+  @media screen and (max-width: 500px) {
+    .perk-overview__toggle {
+      font-size: 18px;
+    }
+
+    .perk-overview__icon-holder {
+      width: 50px;
+    }
+  }
+
+  .perk-overview__grid {
+    padding: 30px 0;
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+    grid-gap: 30px;
+  }
+
+  .fade-enter-active,
+  .fade-leave-active {
+    transition: all .2s ease-in-out;
+    transform: translateY(0);
+    opacity: 1;
+  }
+
+  .fade-enter,
+  .fade-leave-to {
+    transform: translateY(-50px);
+    opacity: 0;
+  }
+}
