+++ src/components/PixiPerkSlot/PixiPerkSlot.vue
@@ -1,0 +1,6 @@
+<template>
+    <div class="slot" ref="slot" :style="cssProps"></div>
+</template>
+
+<script src="./pixiPerkSlot.js"></script>
+<style lang="scss" src="./pixiPerkSlot.scss" scoped></style>
