+++ src/components/UiButton/uiButton.scss
@@ -1,0 +1,37 @@
+@import "../../design/main";
+
+.ui-button {
+  display: inline-block;
+  padding: 4px 35px;
+  color: $color-text;
+  background-color: $color-primary;
+  border: 0 none;
+  text-decoration: none;
+  border-radius: $border-radius-default;
+  font-family: $font-family-default;
+  font-size: $font-size-default;
+  text-transform: uppercase;
+  overflow: hidden;
+  white-space: nowrap;
+  text-align: center;
+
+  .ui-button__icon {
+    display: block;
+    width: 50px;
+
+    img {
+      display: block;
+      max-width: 100%;
+    }
+  }
+
+  &.ui-button--icon {
+    display: flex;
+    align-items: center;
+    padding-left: 15px;
+
+    .ui-button__icon {
+      margin-right: 15px;
+    }
+  }
+}
