+++ src/main.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import App from './App.vue'
+import App from './views/App/App.vue'
 import router from './router'
 import i18n from './i18n'
