+++ src/views/Home/Home.scss
@@ -1,0 +1,20 @@
+@import "../../../public/sprites/surv-css.css";
+@import "../../../public/sprites/kill-css.css";
+
+.language-switch {
+  .flag {
+    margin: 0 0.5rem;
+  }
+
+  border: 1px solid rgba(255, 255, 255, 0.2);
+  background-color: rgba(0, 0, 0, 0.2);
+  padding: 10px 20px;
+  display: inline-block;
+  align-items: center;
+}
+
+@media screen and (max-width: 1055px) {
+  .language-switch {
+    margin-left: 1rem;
+  }
+}
