+++ src/views/NotFound/NotFoundComponent.vue
@@ -1,0 +1,5 @@
+<template>
+    <h1>404: Page not found</h1>
+</template>
+
+<script src="./notFoundComponent.js"></script>
