+++ src/views/NotFound/notFoundComponent.js
@@ -1,0 +1,3 @@
+export default {
+  name: 'NotFoundComponent'
+}
