--- src/views/NotFoundComponent.vue
@@ -1,9 +1,0 @@
-<template>
-    <h1>404: Page not found</h1>
-</template>
-
-<script>
-export default {
-  name: 'NotFoundComponent'
-}
-</script>
